+++ src/index.ts
@@ -1,10 +1,10 @@
-import { Todoist as TodoistV8 } from './v8'
-import * as TodoistV8Types from './v8-types'
+import { Todoist as TodoistV9 } from './v9'
+import * as TodoistV9Types from './v9-types'
 
 // export V8 as the current version
-export const Todoist = TodoistV8
+export const Todoist = TodoistV9
 
 // create backwards compatibility list
-export const v8 = TodoistV8
+export const v9 = TodoistV9
 // export Types as well
-export { TodoistV8Types }
+export { TodoistV9Types }
